Re-activated tests + restructuring complete
