+++ medici_core/src/lib.rs
@@ -45,5 +45,5 @@ pub mod function;
 pub mod marker;
 pub mod service;
 pub mod stm;
-pub mod transaction;
 pub mod storage;
+pub mod transaction;
